Complete all feature related of bill service

+++ Infrastructure/Entities/Order.cs
@@ -14,7 +14,7 @@ namespace Infrastructure.Entities
         [Key]
         public int? OrderID { get; set; }
         public DateTime? OrderDate { get; set; }
-        public int? TotalAmount { get; set; }
+        public Double? TotalAmount { get; set; }
 
         public int? AccountId { get; set; } // khi account này khác null có nghĩa là đơn hàng đã dc xác nhận 
         public bool? IsDone { get; set; }
